Add a build.sh for dell_raid.
This will verify that the utilities the RAID controller needs are in
place. If they are not, it will prompt the user to download what is
needed into the right location in the build cache.

We do it this way because LSI requires that folks agree to their EULA
before using their utilities, and we do not have redistribution rights
to any of their tools.